set fill_predictions test to not run unless asked for
ugh, looks like there's something wonky about running the full fill_predictions code on travis, so for now, just including a manual toggle (test_location) that should be set to "local" if you want to test that. this is also not a bad thing because that function takes a fair amount of time. in the future, goal is to set up the environment for the test running so it's an automatic switch (not a manual one)